This commit is to quell the great thanksgiving rage of 2017. The PI lib was not as robust as I had hoped, causing many unexpected failures, possibley exascerbated(sp?) by a hasty commit on thanksgiving eve.
The changes are mostly checks to see if the PI library tcp
connection has failed and
to make sure the indi driver and client let the user know this
has happened. Unfortunately, after rolling back the hasty
pre thanksgiving changes, I haven't been able to make the
connection fail in the same way. This likely means
it will happen late this weekend. The next step is to see
what causes the failure or failing that, see if it is possible
to recover automatically.